+++ src/App.tsx
@@ -1,32 +1,29 @@
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 
 import { TodoType } from '@components/Todo';
+import TodoAddForm from '@components/TodoAddForm';
 import Todos from '@components/Todos';
 import TodosPresenter from '@presenters/todosPresenter';
 
 const todosPresenter = new TodosPresenter([{ id: 1, title: 'ㅎㅇ' }]);
 
 const App: React.FC = () => {
-  const inputRef = useRef<HTMLInputElement>(null);
   const [todos, setTodos] = useState<TodoType[]>(todosPresenter.getTodos());
 
   const handleClickDeleteTodo = (todoId) => {
     todosPresenter.delete(todoId, setTodos);
   };
 
-  const handleSubmit = () => {
-    if (!inputRef.current) return;
-    const title = inputRef.current.value;
+  const handleClickAddTodo = (title) => {
     todosPresenter.add(title, setTodos);
   };
 
   return (
-    <form onSubmit={handleSubmit}>
+    <>
       <h1>Hello World</h1>
-      <input ref={inputRef} type="text" />
-      <button type="submit">생성</button>
+      <TodoAddForm onTodoAdd={handleClickAddTodo} />
       <Todos todos={todos} onClickDeleteTodo={handleClickDeleteTodo} />
-    </form>
+    </>
   );
 };
